refactor: rename SegmentControl to SegmentControl_Tight
