+++ src/components/SegmentControl/index.ts
@@ -1,2 +1,2 @@
-export { default } from "./SegmentControl";
+export { default } from "./SegmentControl_Tight";
